+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\SubCategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\ChildCategory;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
@@ -100,6 +101,11 @@ class SubCategoryController extends Controller
     {
         //
         $subcategory = SubCategory::findOrFail($id);
+        $childCategory = ChildCategory::where('sub_category_id', $subcategory->id)->count();
+        if($childCategory > 0){
+            return response(['status'=>'error', 'message'=>'This item contains sub items for delete, you need to
+            delete sub items first!!!']);
+        }
         $subcategory->delete();
         return response(['status'=>'success', 'Deleted Successfully!!']);
         return redirect()->route('admin.sub-category.index');
